Extract a carmon network

get_network() extracts the reconstructed (and selected) network from a carmon object.

Usage

get_network(carmon_obj, labels = NULL)

Arguments

carmon_obj

The object of S3 class carmon or of S3 class carmon_rec.

labels

Optional. Used only when carmon_obj is of the S3 class carmon_rec, a vector of strings containing all the variable names.

Value

get_network() returns the selected network, in the form of an object of class igraph. If using an object of class carmon_rec, nodes will have no labels, unless provided through labels.

Examples

data(multi_omics_small)
c_obj <- carmon(multi_omics_small,
    net_method = "correlation",
    cor_quant = 0.25, analysis = FALSE, plot = FALSE,
    verbose = FALSE
)
network <- get_network(c_obj)