Display the table of results of the consensus centrality analysis of the carmon network

Display the table of results of the consensus centrality analysis of the carmon network

centrality_report(carmon_obj)

Arguments

carmon_obj

The object of S3 class carmon, possibly already resulting from a full carmon() run, or at least undergone centrality analysis through compute_centrality().

Value

centrality_report() returns the centrality report of the object of S3 class carmon given as an input, in the form of an R data frame. The rows correspond to the nodes identified to be central by the analysis, and they are ordered based on how large is the consensus among the different measures. The data frame has 6 columns:
candidate, the name of the central node;
degree, the degree centrality of the node;
betweenness, the betweenness centrality of the node;
closeness, the closeness centrality of the node;
eigenvector, the eigenvector centrality of the node; and
central for, the a string reporting the first letter of all the measures according to which the node is central for.

Examples

# Let's build and analyse a carmon network:
data(multi_omics_small)
c_obj <- carmon(multi_omics_small,
    net_method = "correlation",
    cor_quant = 0.25, analysis = TRUE, plot = FALSE,
    # analysis is already TRUE by default
    verbose = FALSE
)
# To display the table of the results of the centrality analysis:
centrality_report(c_obj)
#>          candidate degree betweenness closeness eigenvector central for
#> BC004004  BC004004   0.5*       0.138   0.6071*          1*         dce
#> P4ha1        P4ha1 0.2222     0.2308*    0.5152      0.1935           b